+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Add support for Level 3 summary parameters, shippingAmount, discountAmount, and shipsFromPostalCode
 * Add support for transaction line items
 * Add support for tagged evidence in DisputeGateway#addTextEvidence (Beta release)